Move to its own file. rename handle_snapshot_request to serve_snapshot_request
